refactor: extract skill promotion support
